+++ src/main/java/com/se/authserver/v1/common/infra/security/util/SecurityContextUtil.java
@@ -1,5 +1,6 @@
 package com.se.authserver.v1.common.infra.security.util;
 
+import com.se.authserver.v1.common.domain.exception.UnauthenticatedException;
 import java.util.Set;
 import lombok.AccessLevel;
 import lombok.NoArgsConstructor;
@@ -10,6 +11,8 @@ import org.springframework.security.core.context.SecurityContextHolder;
 public class SecurityContextUtil {
 
   public static Long getCurrentAccountId() {
+    if (SecurityContextHolder.getContext().getAuthentication() == null)
+      throw new UnauthenticatedException("인증 정보가 존재하지 않습니다");
     return Long.parseLong(SecurityContextHolder.getContext().getAuthentication().getName());
   }
